add unread tray variant & make it customisable
