+++ src/main/events.ts
@@ -11,4 +11,5 @@ import { UserAssetType } from "./userAssets";
 export const AppEvents = new EventEmitter<{
     appLoaded: [];
     userAssetChanged: [UserAssetType];
+    setTrayVariant: ["tray" | "trayUnread"];
 }>();
