+++ src/main/mainWindow.ts
@@ -8,15 +8,12 @@ import {
     app,
     BrowserWindow,
     BrowserWindowConstructorOptions,
-    dialog,
     Menu,
     MenuItemConstructorOptions,
     nativeTheme,
     screen,
-    session,
-    Tray
+    session
 } from "electron";
-import { rm } from "fs/promises";
 import { join } from "path";
 import { IpcCommands, IpcEvents } from "shared/IpcEvents";
 import { isTruthy } from "shared/utils/guards";
@@ -25,28 +22,19 @@ import type { SettingsStore } from "shared/utils/SettingsStore";
 
 import { createAboutWindow } from "./about";
 import { initArRPC } from "./arrpc";
-import {
-    BrowserUserAgent,
-    DATA_DIR,
-    DEFAULT_HEIGHT,
-    DEFAULT_WIDTH,
-    MessageBoxChoice,
-    MIN_HEIGHT,
-    MIN_WIDTH,
-    VENCORD_FILES_DIR
-} from "./constants";
+import { BrowserUserAgent, DEFAULT_HEIGHT, DEFAULT_WIDTH, MIN_HEIGHT, MIN_WIDTH, VENCORD_FILES_DIR } from "./constants";
 import { AppEvents } from "./events";
 import { darwinURL } from "./index";
 import { sendRendererCommand } from "./ipcCommands";
 import { Settings, State, VencordSettings } from "./settings";
 import { createSplashWindow, updateSplashMessage } from "./splash";
-import { resolveAssetPath } from "./userAssets";
+import { destroyTray, initTray } from "./tray";
+import { clearData } from "./utils/clearData";
 import { makeLinksOpenExternally } from "./utils/makeLinksOpenExternally";
 import { applyDeckKeyboardFix, askToApplySteamLayout, isDeckGameMode } from "./utils/steamOS";
 import { downloadVencordFiles, ensureVencordFiles } from "./utils/vencordLoader";
 
 let isQuitting = false;
-let tray: Tray;
 
 applyDeckKeyboardFix();
 
@@ -76,90 +64,6 @@ function makeSettingsListenerHelpers<O extends object>(o: SettingsStore<O>) {
 
 const [addSettingsListener, removeSettingsListeners] = makeSettingsListenerHelpers(Settings);
 const [addVencordSettingsListener, removeVencordSettingsListeners] = makeSettingsListenerHelpers(VencordSettings);
-
-AppEvents.on("userAssetChanged", async asset => {
-    if (asset === "tray" && tray) {
-        tray.setImage(await resolveAssetPath("tray"));
-    }
-});
-
-async function initTray(win: BrowserWindow) {
-    const onTrayClick = () => {
-        if (Settings.store.clickTrayToShowHide && win.isVisible()) win.hide();
-        else win.show();
-    };
-    const trayMenu = Menu.buildFromTemplate([
-        {
-            label: "Open",
-            click() {
-                win.show();
-            }
-        },
-        {
-            label: "About",
-            click: createAboutWindow
-        },
-        {
-            label: "Repair Vencord",
-            async click() {
-                await downloadVencordFiles();
-                app.relaunch();
-                app.quit();
-            }
-        },
-        {
-            label: "Reset Vesktop",
-            async click() {
-                await clearData(win);
-            }
-        },
-        {
-            type: "separator"
-        },
-        {
-            label: "Restart",
-            click() {
-                app.relaunch();
-                app.quit();
-            }
-        },
-        {
-            label: "Quit",
-            click() {
-                isQuitting = true;
-                app.quit();
-            }
-        }
-    ]);
-
-    tray = new Tray(await resolveAssetPath("tray"));
-    tray.setToolTip("Vesktop");
-    tray.setContextMenu(trayMenu);
-    tray.on("click", onTrayClick);
-}
-
-async function clearData(win: BrowserWindow) {
-    const { response } = await dialog.showMessageBox(win, {
-        message: "Are you sure you want to reset Vesktop?",
-        detail: "This will log you out, clear caches and reset all your settings!\n\nVesktop will automatically restart after this operation.",
-        buttons: ["Yes", "No"],
-        cancelId: MessageBoxChoice.Cancel,
-        defaultId: MessageBoxChoice.Default,
-        type: "warning"
-    });
-
-    if (response === MessageBoxChoice.Cancel) return;
-
-    win.close();
-
-    await win.webContents.session.clearStorageData();
-    await win.webContents.session.clearCache();
-    await win.webContents.session.clearCodeCaches({});
-    await rm(DATA_DIR, { force: true, recursive: true });
-
-    app.relaunch();
-    app.quit();
-}
 
 type MenuItemList = Array<MenuItemConstructorOptions | false>;
 
@@ -339,8 +243,8 @@ function initWindowBoundsListeners(win: BrowserWindow) {
 
 function initSettingsListeners(win: BrowserWindow) {
     addSettingsListener("tray", enable => {
-        if (enable) initTray(win);
-        else tray?.destroy();
+        if (enable) initTray(win, q => (isQuitting = q));
+        else destroyTray();
     });
 
     addSettingsListener("disableMinSize", disable => {
@@ -482,7 +386,9 @@ function createMainWindow() {
     });
 
     initWindowBoundsListeners(win);
-    if (!isDeckGameMode && (Settings.store.tray ?? true) && process.platform !== "darwin") initTray(win);
+    if (!isDeckGameMode && (Settings.store.tray ?? true) && process.platform !== "darwin")
+        initTray(win, q => (isQuitting = q));
+
     initMenuBar(win);
     makeLinksOpenExternally(win);
     initSettingsListeners(win);
